rearrange gateway2 logic
